[Rename] Rename weather directory
