+++ app/src/main/java/com/twoday/todaytrip/weatherApi/WeatherDTO.kt
@@ -1,4 +1,4 @@
-package com.twoday.todaytrip.weather
+package com.twoday.todaytrip.weatherApi
 
 import com.google.gson.annotations.SerializedName
